Add prefix parameter to the sudoers definition
The order of the sudo rules matters, and the $title may not start with
numbers since it's used as-is for the sudo users, cmds, etc. aliases. It
was possible to add 'a_', 'z_', etc. to current titles, but it's cleaner
to force ordering by relying on an optional separate parameter.